refactor:Separate window action related functions
